+++ experimental/packages/opentelemetry-sdk-node/test/sdk.test.ts
@@ -696,10 +696,9 @@ describe('Node SDK', () => {
       sdk.shutdown();
     });
 
-    it('should configure service instance id with random UUID with OTEL_NODE_JS_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID env var', async () => {
+    it('should configure service instance id with random UUID with OTEL_NODE_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID env var', async () => {
       process.env.OTEL_RESOURCE_ATTRIBUTES = 'service.name=my-service';
-      process.env.OTEL_NODE_JS_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID =
-        'true';
+      process.env.OTEL_NODE_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID = 'true';
       const sdk = new NodeSDK();
 
       sdk.start();
@@ -711,15 +710,14 @@ describe('Node SDK', () => {
         36
       );
       delete process.env.OTEL_RESOURCE_ATTRIBUTES;
-      delete process.env.OTEL_NODE_JS_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID;
+      delete process.env.OTEL_NODE_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID;
       await sdk.shutdown();
     });
 
-    it('should configure service instance id via OTEL_RESOURCE_ATTRIBUTES env var even with OTEL_NODE_JS_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID env var', async () => {
+    it('should configure service instance id via OTEL_RESOURCE_ATTRIBUTES env var even with OTEL_NODE_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID env var', async () => {
       process.env.OTEL_RESOURCE_ATTRIBUTES =
         'service.instance.id=627cc493,service.name=my-service';
-      process.env.OTEL_NODE_JS_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID =
-        'true';
+      process.env.OTEL_NODE_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID = 'true';
       const sdk = new NodeSDK();
 
       sdk.start();
@@ -731,14 +729,13 @@ describe('Node SDK', () => {
         instanceId: '627cc493',
       });
       delete process.env.OTEL_RESOURCE_ATTRIBUTES;
-      delete process.env.OTEL_NODE_JS_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID;
+      delete process.env.OTEL_NODE_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID;
       sdk.shutdown();
     });
 
-    it('should not configure service instance id with no value for it on OTEL_RESOURCE_ATTRIBUTES env var and  OTEL_NODE_JS_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID env var as false', async () => {
+    it('should not configure service instance id with no value for it on OTEL_RESOURCE_ATTRIBUTES env var and  OTEL_NODE_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID env var as false', async () => {
       process.env.OTEL_RESOURCE_ATTRIBUTES = 'service.name=my-service';
-      process.env.OTEL_NODE_JS_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID =
-        'false';
+      process.env.OTEL_NODE_EXPERIMENTAL_DEFAULT_SERVICE_INSTANCE_ID = 'false';
       const sdk = new NodeSDK();
 
       sdk.start();